requirements.txt
Flask==3.0.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
Flask-Migrate==4.0.5
alembic==1.13.1

# Auth
Flask-Login==0.6.3
itsdangerous==2.1.2

# Forms & validation
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.1.0.post1

# Slugs
python-slugify==8.0.4
text-unidecode==1.3

# Environment variables
python-dotenv==1.0.1

# File handling / security
Werkzeug==3.0.1

gunicorn==21.2.0
